refactor(frontend): Move blocks to subfolder
