+++ frontend/src/components/common/rich-text-editor/blocks/TableWrapper.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react'
-import { EditorContext } from './RichTextEditor'
+import { EditorContext } from '../RichTextEditor'
 
 export const TableWrapper = (props: any) => {
   const { editorState } = useContext(EditorContext)
