+++ frontend/src/components/common/rich-text-editor/blocks/index.ts
@@ -1,0 +1,2 @@
+export { ImageBlock } from './ImageBlock'
+export { TableWrapper } from './TableWrapper'
